# core: js errors as unions vs tuples to reduce allocs