This is an op-layer optimization follow-up to #9843.

It shaves off `~350ns/op` of baseline overhead, making them over 2x faster by reducing allocs.

Instead of returning JS results as tuples of `(Option<T>, Option<Err>)` it returns them as a conceptual union type, avoiding allocating the wrapping array. We then use a "special" key name to recognize errors.

So far I've picked the error-key to be `$err_class_name`, but this can be easily changed or adjusted.

## Results

Op-baseline benches:
```
Before:
test bench_op_nop     ... bench:     321,740 ns/iter (+/- 39,892)
test bench_op_pi_bin  ... bench:     786,395 ns/iter (+/- 23,765)
test bench_op_pi_json ... bench:     681,285 ns/iter (+/- 26,579)

After:
test bench_op_nop     ... bench:     318,910 ns/iter (+/- 16,233)
test bench_op_pi_bin  ... bench:     402,301 ns/iter (+/- 28,707)
test bench_op_pi_json ... bench:     294,627 ns/iter (+/- 14,079)
```